-fix: updated time according to new game timeline

+++ src/main/java/gamein2022/backend/dashboard/web/dto/result/TimeResultDTO.java
@@ -12,7 +12,7 @@ public class TimeResultDTO implements BaseResultDTO {
     private Long day;
     private Long month;
     private Long year;
-    private Long era;
+    private Byte era;
     private Long secondOfDate;
 
 }
